server/requirements.txt
openenv-core
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
torch>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
